Update README.md
At the beginning of the program, header files mbed.h and rtos.h are included in the program, and the serial PC interface is
defined. EntrySwitch, ExitSwitch, and StartButton are assigned to PC_0, PC_1, and PC_3,respectively. LockEntryBarrier output when set locks the entry barrier so that it does not open
and it is assigned to PC_2. The car park capacity is set to 100 cars. Inside the main program message CLOSED is displayed since the car park is closed at the beginning of the program. Threads
CarParkandDisplayare started. These threadswait forevent flags 1 and 2 to be set, respectively,
before they can continue. These flags are set when the StartButton is pressed (i.e., StartButton is
at logic LOW). Setting the flags starts the two threads. Thread Display displays the spaces available in the car park at screen coordinate (0, 9). Thread CarPark increases or decreases the space
count depending on whether the cars are entering or leaving the car park. When a car enters the
car park variable Spaces is decremented by one. Similarly, when a car leaves the car park variableSpaces isincremented by one. Theprogrammakes sure that the space countis not above the
car park capacity or below zero (this should never happen in practice).
Note that this project could have been implemented without using multitasking, but here
multitasking is used for demonstration purposes.